currency-exchange-service/src/main/java/com/currency/currencyexchangeservice/controller/CurrencyExchangeController.java
package com.currency.currencyexchangeservice.controller;

import com.currency.currencyexchangeservice.entity.CurrencyExchangeRate;
import com.currency.currencyexchangeservice.service.ExchangeService;
import org.springframework.batch.core.BatchStatus;
import org.springframework.batch.core.JobParametersInvalidException;
import org.springframework.batch.core.repository.JobExecutionAlreadyRunningException;
import org.springframework.batch.core.repository.JobInstanceAlreadyCompleteException;
import org.springframework.batch.core.repository.JobRestartException;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.math.BigDecimal;

@RestController
@RequestMapping("/api")
public class CurrencyExchangeController {

    @Autowired
    private ExchangeService exchangeService;

    @GetMapping("/from/{from}/to/{to}")
    public ResponseEntity<CurrencyExchangeRate> getCurrencyExchangeRate(@PathVariable String from, @PathVariable String to){
        try {
            CurrencyExchangeRate result = exchangeService.findRate(from,to);
            return new ResponseEntity<>(result, HttpStatus.OK);
        }catch (Exception e){
            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
        }
    }

    @PostMapping("/rate")
    public ResponseEntity<CurrencyExchangeRate> addCurrencyExchangeRate(@RequestBody CurrencyExchangeRate currencyExchangeRate){
        try {
            CurrencyExchangeRate result = exchangeService.addRate(currencyExchangeRate);
            return new ResponseEntity<>(result, HttpStatus.OK);
        }catch (Exception e){
            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
        }

    }

    @GetMapping("/load")
    public BatchStatus load() throws JobInstanceAlreadyCompleteException, JobExecutionAlreadyRunningException, JobParametersInvalidException, JobRestartException {
        return exchangeService.loadDataFromCSVtoDB();
    }

}
